refactor: HTTP 로깅 내용을 캡슐화
