+++ backend/src/main/java/coursepick/coursepick/logging/HttpLoggingFilter.java
@@ -11,8 +11,6 @@ import org.springframework.web.util.ContentCachingRequestWrapper;
 import org.springframework.web.util.ContentCachingResponseWrapper;
 
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.util.Enumeration;
 
 @Slf4j
 @Component
@@ -27,49 +25,9 @@ public class HttpLoggingFilter extends OncePerRequestFilter {
         filterChain.doFilter(requestWrapper, responseWrapper);
         long duration = System.currentTimeMillis() - startTime;
 
-        doLog(requestWrapper, responseWrapper, duration);
+        String content = LogContentCreator.http(requestWrapper, responseWrapper, duration);
+        log.info(content);
+
         responseWrapper.copyBodyToResponse();
-    }
-
-    private static void doLog(ContentCachingRequestWrapper request, ContentCachingResponseWrapper response, long duration) {
-        String method = request.getMethod();
-        String uri = extractUriWithQueryString(request);
-        String headers = extractHeaderString(request);
-        String requestBody = new String(request.getContentAsByteArray(), StandardCharsets.UTF_8).strip();
-        int status = response.getStatus();
-        String responseBody = new String(response.getContentAsByteArray(), StandardCharsets.UTF_8).strip();
-
-        String logContent = createLogContent(method, uri, duration, headers, requestBody, status, responseBody);
-        log.info(logContent);
-    }
-
-    private static String extractUriWithQueryString(ContentCachingRequestWrapper request) {
-        String requestURI = request.getRequestURI();
-        String queryString = request.getQueryString();
-
-        if (queryString != null) {
-            requestURI += "?" + queryString;
-        }
-        return requestURI;
-    }
-
-    private static String extractHeaderString(ContentCachingRequestWrapper request) {
-        StringBuilder headers = new StringBuilder();
-        Enumeration<String> headerNames = request.getHeaderNames();
-        while (headerNames.hasMoreElements()) {
-            String headerName = headerNames.nextElement();
-            headers.append("[").append(headerName).append(": ").append(request.getHeader(headerName)).append("] ");
-        }
-        return headers.toString();
-    }
-
-    private static String createLogContent(String method, String uri, long duration, String headers, String requestBody, int status, String responseBody) {
-        return """
-                [HTTP] %s %s (%dms)
-                    Req Headers: %s
-                    Req Body: %s
-                    Res Status: %d
-                    Res Body: %s
-                """.formatted(method, uri, duration, headers, requestBody, status, responseBody);
     }
 }
